Reorganise, remove old functions
